Montgomery Fleets Data Analysis
This dataset comprises fleet data for Montgomery Fleets, sourced from Kaggle. Prior to analysis, we diligently cleaned the dataset by eliminating empty and redundant rows, as well as correcting spelling errors and erroneous data types. Once the data was refined, we conducted an in-depth analysis utilizing pivot tables and various statistical methods, the details of which are outlined in the accompanying files. The results of our analysis offer valuable insights into Montgomery Fleets' operations and performance.